requirements.txt
asyncpg
aiogram==3.12.0
APScheduler==3.10.4
python-decouple==3.8
git+https://github.com/weatherapicom/python.git#egg=weatherapipython
pytz
SQLAlchemy==2.0.32
redis==5.0.8